1. Create a custom environment "highway-fast-v1"
"highway-fast-v1" shifts all random vehicles, except for controlled agents, backward in the longitudinal direction.

In the original settings, all the other vehicles spawn only in front of controlled agents. This would lead to an unexpected learning result that the agent would linger (driving a circle in place) to avoid conflicts with other vehicles.